Implementation of the PCB types builder functions.
For the moment, a PCB is a gas volume surrounded by two PCB plates made
of copper. All the PCBs are built such that the bending side faces the
IP (z>0). The json takes care of the necessary rotations when the slats
are placed on the half-chambers so that we match the mapping convention.