Attempt (failed to far) to solve the Root vs Cuda vs C++17 issue.
See also https://github.com/AliceO2Group/AliceO2/pull/2186

The idea here would be to create a "clone" of the target(s) Cuda depends
on, and relax the C++17 requirement on those clones.

Then the Cuda targets would depend on those clones instead of the
original targets.

Failed so far as some part of the include directories does not seem to
be passed on to nvcc ...

To be continued ...