Update of the MCH Raw Data Decoder (in particular the UserLogic one).
We're introducing in this commit :

-   a new class for the Electronic Mapping
      : the FeeLinkId, which is a pair of (FeeId,LinkId) where FeeId is (or
      will be) ranging from 0 to 61, i.e. to the number of endpoints we have
      for MCH (31 CRUs, each with 2 endpoints), and LinkId from 0 to 11.
-   the electronic mapping for CH8L (needed for DE819 used as a guinea pig
      for various tests)
-   a revamped implementation of the UL decoder

We're removing :

- RawDataHeaderHandler which won't be used in the long term. Currently
we still need to "patch" the buffer to tweak the RDH before passing it
to the PageDecoder, but that is temporary and can be done without
introducing a type (by just patching the raw buffer directly).

A note on the UserLogic Decoder implementation. After some attempts to
use proper State Machine libraries (Boost MSM or boost experimental SML) the
conclusion was that, while they provided somewhat cleaner code, they
suffered from some defects :

-   MSM is (way) too slow to compile
-   SML is faster to compile and seemed
     promising, but we've never been able to completely solve a bug we
     encountered when using it, so had to back off

So we dropped the ball and got back to a hand-made coding of a
state-machine-like decoder. It's not as cleaner than either of the two
libraries above, but ... it works...