[MCH] implemented new definition of digit time
The new definition combines the SAMPA time stamp, the SAMPA bunch crossing counter,
and the orbit number of the HBFrame containing the digit.

The introduction of this new time definition requires adding the bunch crossing counter to the
SampaCluster strcture, and therefore to change the Elink decoders as well as some of the unity tests.

The new definition does not change the size of the Digit structure.

[MCH] added two reserved bits in the hit time structure

[MCH] fixed bug in custom mapping loader code

[MCH] moved digit time structure inside Digit class definition

The time parameter has been also moved to the last position in the Digit constructor,
and the old "double" parameter has been removed.

This modification required several changes in the Raw encoder and simulation parts.

[MCH] renamed timestamp as sampaTime in SampaCluster struct

[MCH] fixed default bunch crossing value in SampaCluster testing code

[MCH] update encoder tests to include bunchCrossing